requirements.txt
altair<5 # altair==4.0.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
boto3==1.34.123
botocore==1.34.123
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.6
entrypoints==0.4
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.14.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.3.0
protobuf==3.20.3
pyarrow==16.1.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.12.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.28.3
watchdog==4.0.1
zipp==3.19.2
